+++ core/src/main/scala/astraea/spark/rasterframes/expressions/mapalgebra/Add.scala
@@ -22,17 +22,24 @@
 package astraea.spark.rasterframes.expressions.mapalgebra
 
 import astraea.spark.rasterframes._
+import astraea.spark.rasterframes.expressions.{BinaryLocalRasterOp, BinaryRasterOp}
 import geotrellis.raster.Tile
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
+import org.apache.spark.sql.functions.lit
 import org.apache.spark.sql.{Column, TypedColumn}
 
 /** Performs cell-wise addition between two tiles. */
-case class Add(left: Expression, right: Expression) extends BinaryRasterOp with CodegenFallback {
+case class Add(left: Expression, right: Expression) extends BinaryLocalRasterOp with CodegenFallback {
   override val nodeName: String = "local_add"
   override protected def op(left: Tile, right: Tile): Tile = left.localAdd(right)
+  override protected def op(left: Tile,  right: Double): Tile = left.localAdd(right)
+  override protected def op(left: Tile, right: Int): Tile = left.localAdd(right)
 }
 object Add {
   def apply(left: Column, right: Column): TypedColumn[Any, Tile] =
     new Column(Add(left.expr, right.expr)).as[Tile]
+
+  def apply[N: Numeric](tile: Column, value: N): TypedColumn[Any, Tile] =
+    new Column(new Add(tile.expr, lit(value).expr)).as[Tile]
 }