+++ core/src/main/scala/astraea/spark/rasterframes/expressions/mapalgebra/Subtract.scala
@@ -21,17 +21,24 @@
 
 package astraea.spark.rasterframes.expressions.mapalgebra
 import astraea.spark.rasterframes._
+import astraea.spark.rasterframes.expressions.{BinaryLocalRasterOp, BinaryRasterOp}
 import geotrellis.raster.Tile
 import org.apache.spark.sql.{Column, TypedColumn}
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
+import org.apache.spark.sql.functions.lit
 
 /** Performs cell-wise subtraction between two tiles. */
-case class Subtract(left: Expression, right: Expression) extends BinaryRasterOp with CodegenFallback {
+case class Subtract(left: Expression, right: Expression) extends BinaryLocalRasterOp with CodegenFallback {
   override val nodeName: String = "local_subtract"
   override protected def op(left: Tile, right: Tile): Tile = left.localSubtract(right)
+  override protected def op(left: Tile, right: Double): Tile = left.localSubtract(right)
+  override protected def op(left: Tile, right: Int): Tile = left.localSubtract(right)
 }
 object Subtract {
   def apply(left: Column, right: Column): TypedColumn[Any, Tile] =
     new Column(Subtract(left.expr, right.expr)).as[Tile]
+
+  def apply[N: Numeric](tile: Column, value: N): TypedColumn[Any, Tile] =
+    new Column(new Subtract(tile.expr, lit(value).expr)).as[Tile]
 }