+++ pyrasterframes/python/tests/PyRasterFramesTests.py
@@ -190,9 +190,9 @@ class RasterFunctionsTest(unittest.TestCase):
         mask_value = 4
 
         rf1 = self.rf.select(self.rf.tile,
-                             local_multiply_scalar_int(
+                             local_multiply(
                                  convert_cell_type(
-                                     local_greater_scalar_int(self.rf.tile, 25000),
+                                     local_greater(self.rf.tile, 25000),
                                      "uint8"),
                                   mask_value).alias('mask'))
         rf2 = rf1.select(rf1.tile, mask_by_value(rf1.tile, rf1.mask, lit(mask_value)).alias('masked'))